add suggestLocation form

+++ workspace/assignment_sprint1/sprint1/models.py
@@ -75,8 +75,8 @@ class Tag(models.Model):
 
 class LocationSuggestion(models.Model):
     locationName = models.CharField(max_length=254, null=True)
-    latitude = models.IntegerField(null=True)
-    longitude = models.IntegerField(null=True)
+    latitude = models.FloatField(null=True)
+    longitude = models.FloatField(null=True)
     locationAddress = models.CharField(max_length=254, null=True)
     locationBio = models.TextField(max_length=511, null=True)
     locationType = models.IntegerField(null=True)
